fix(backups): move pure helpers out of Next.js page files into utils modules
Next.js rejects named exports from page files that aren't standard page
exports. Extract formatSnapshotLabel/flattenTree/getNodeIcon and
getExportFilename/getDownloadButtonLabel into sibling utils.ts files.
